# src/main/resources/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: BiniTech PDV API
  description: API do Sistema Frente de Caixa (PDV) da BiniTech
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Servidor local de desenvolvimento

tags:
  - name: Auth
    description: Autenticação e registro de usuários
  - name: Products
    description: Operações de cadastro e consulta de produtos
  - name: Sales
    description: Operações de venda e relatórios

paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Fazer login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
        '400':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/auth/register:
    post:
      tags: [Auth]
      summary: Registrar novo usuário (somente ADMIN)
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access token usando refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequestDTO'
      responses:
        '200':
          description: Token renovado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
        '400':
          description: Refresh token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Fazer logout e invalidar token
      operationId: logout
      responses:
        '204':
          description: Logout realizado com sucesso

  /api/products:
    get:
      tags: [Products]
      summary: Listar todos os produtos
      operationId: listProducts
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
    post:
      tags: [Products]
      summary: Cadastrar um novo produto
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDTO'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/products/{id}:
    get:
      tags: [Products]
      summary: Buscar produto por ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
    put:
      tags: [Products]
      summary: Atualizar um produto existente
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDTO'
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
    delete:
      tags: [Products]
      summary: Desativar um produto
      operationId: deactivateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Produto desativado com sucesso
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/products/barcode/{barcode}:
    get:
      tags: [Products]
      summary: Buscar produto por código de barras (leitura rápida)
      operationId: getProductByBarcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/sales:
    post:
      tags: [Sales]
      summary: Registrar uma nova venda
      operationId: createSale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleDTO'
      responses:
        '201':
          description: Venda registrada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleDTO'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
    get:
      tags: [Sales]
      summary: Listar vendas com filtro por data
      operationId: listSales
      parameters:
        - name: date
          in: query
          required: false
          description: Data específica (formato yyyy-MM-dd) para extrato do dia
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          required: false
          description: Data inicial do período (formato yyyy-MM-dd)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          description: Data final do período (formato yyyy-MM-dd)
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        '200':
          description: Lista de vendas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaleDTO'

  /api/sales/{id}:
    get:
      tags: [Sales]
      summary: Buscar venda por ID
      operationId: getSaleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Venda encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleDTO'
        '404':
          description: Venda não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/sales/{id}/mark-paid:
    patch:
      tags: [Sales]
      summary: Marcar venda crediário como paga
      operationId: markSaleAsPaid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Venda marcada como paga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaleDTO'
        '404':
          description: Venda não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'

  /api/sales/debtors:
    get:
      tags: [Sales]
      summary: Listar vendas com crediário não pagas (devedores)
      operationId: listDebtors
      responses:
        '200':
          description: Lista de vendas crediário não pagas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SaleDTO'

components:
  schemas:
    LoginRequestDTO:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        tenantId:
          type: string

    RegisterRequestDTO:
      type: object
      required:
        - username
        - password
        - role
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        tenantId:
          type: string
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - ADMIN
            - TENANT_ADMIN
            - OPERATOR

    RefreshTokenRequestDTO:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    AuthResponseDTO:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        username:
          type: string
        role:
          type: string
        tenantId:
          type: string

    ProductDTO:
      type: object
      properties:
        id:
          type: string
        barcode:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        costPrice:
          type: number
          format: double
        stockQuantity:
          type: integer
        active:
          type: boolean
        category:
          type: string
        userId:
          type: string
          readOnly: true

    CreateProductDTO:
      type: object
      required:
        - barcode
        - description
        - price
        - stockQuantity
      properties:
        barcode:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        price:
          type: number
          format: double
          minimum: 0.01
        costPrice:
          type: number
          format: double
          minimum: 0
        stockQuantity:
          type: integer
          minimum: 0
        category:
          type: string
        active:
          type: boolean

    SaleDTO:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SaleItemDTO'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentDTO'
        totalAmount:
          type: number
          format: double
        totalCost:
          type: number
          format: double
        totalPaid:
          type: number
          format: double
        change:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
        userId:
          type: string
          readOnly: true
        customerName:
          type: string
        customerPhone:
          type: string
        paid:
          type: boolean

    SaleItemDTO:
      type: object
      properties:
        productId:
          type: string
        productDescription:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double
        costPrice:
          type: number
          format: double
        subtotal:
          type: number
          format: double

    CreateSaleDTO:
      type: object
      required:
        - items
        - payments
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateSaleItemDTO'
        payments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PaymentDTO'
        customerName:
          type: string
        customerPhone:
          type: string
        skipStockValidation:
          type: boolean

    CreateSaleItemDTO:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1

    PaymentDTO:
      type: object
      required:
        - method
        - amount
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethodEnum'
        amount:
          type: number
          format: double
          minimum: 0.01

    PaymentMethodEnum:
      type: string
      enum:
        - CASH
        - CREDIT_CARD
        - DEBIT_CARD
        - PIX
        - CREDIARIO

    ErrorDTO:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time